requirements.txt
# Python 3.14.3
python-dotenv == 1.2.2
